+++ examples/default-vpc-iam/outputs.tf
@@ -1,0 +1,4 @@
+output "server_url" {
+  value       = module.rundeck.server_url
+  description = "HTTPS endpoint of Rundeck host"
+}