+++ examples/default-vpc/main.tf
@@ -42,25 +42,3 @@ data "aws_subnet" "default" {
   availability_zone = data.aws_availability_zones.available.names[0]
   vpc_id            = data.aws_vpc.default.id
 }
-
-# resource "aws_iam_policy" "rundeck" {
-#   name        = "rundeck-io-policy"
-#   description = "Policy for EC2 Rundeck"
-#   policy = jsonencode({
-#     Version = "2012-10-17",
-#     Statement = [{
-#       Action = [
-#         "s3:Get*",
-#         "s3:List*"
-#       ],
-#       Effect   = "Allow",
-#       Resource = "*"
-#       },
-#       {
-#         "Effect" : "Allow",
-#         "Action" : ["ec2:Describe*"],
-#         "Resource" : "*"
-#       }
-#     ]
-#   })
-# }